+++ packages/eas-cli/src/prompts.ts
@@ -45,7 +45,10 @@ export async function confirmAsync(
 export async function selectAsync<T>(
   message: string,
   choices: ExpoChoice<T>[],
-  options?: Options
+  config?: {
+    options?: Options;
+    warningMessageForDisabledEntries?: string;
+  }
 ): Promise<T> {
   const { value } = await promptAsync(
     {
@@ -53,8 +56,9 @@ export async function selectAsync<T>(
       choices,
       name: 'value',
       type: 'select',
+      warn: config?.warningMessageForDisabledEntries,
     },
-    options
+    config?.options ?? {}
   );
   return value ?? null;
 }